requirements.txt
# Core API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ML / Search
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
numpy>=1.24.0

# Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# LLM
google-genai>=1.0.0

# Data
pandas>=2.0.0
openpyxl>=3.1.0
